Add helper functions to return most significant persoanlity trait
I created a helper function that takes the Watson API result as
input and iterates through the 30 personality traits in the Big 5
categories. It then compares each of the traits and returns the
trait with the highest percentile score. This trait will be fed to a function that will return the appropriate Pokemon type.