Add successful connection to MongoDB cloud database
I added a connection to my MongoDB Atlas cloud database in app.js.
The next move will be trying to either ecport the db into a separate
file that I can call in each instance it is used, or to store a document
with my User schema into the database.